+++ src/pipecat/services/stt_service.py
@@ -312,6 +312,8 @@ class WebsocketSTTService(STTService, WebsocketService):
         """
         STTService.__init__(self, **kwargs)
         WebsocketService.__init__(self, reconnect_on_error=reconnect_on_error, **kwargs)
+        self._register_event_handler("on_connected")
+        self._register_event_handler("on_disconnected")
         self._register_event_handler("on_connection_error")
 
     async def _report_error(self, error: ErrorFrame):